ESWE-830_manage_application moved getApplicationProgress to correct service
