+++ server/data/jobApplicationApi/getApplicationProgressResponse.ts
