+++ server/data/jobApplicationApi/jobApplicationApi.ts
@@ -1,4 +1,7 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+
 import config from '../../config'
+import ApplicationStatusValue from '../../enums/applicationStatusValue'
 import RestClient from '../restClient'
 
 // const BASE_URL = '/applications'
@@ -16,6 +19,24 @@ export default class JobApplicationApiClient {
 
   async getClosedApplications(offenderNo: string) {
     return mockClosedApplications[offenderNo] ? mockClosedApplications[offenderNo] : mockClosedApplications.default
+  }
+
+  async getApplicationProgress(offenderNo: string, jobId: string) {
+    return [
+      {
+        status: ApplicationStatusValue.APPLICATION_MADE,
+        createdByName: 'TE£ST_USER',
+        createdByDateTime: new Date().toISOString(),
+        notes: '',
+      },
+      {
+        status: ApplicationStatusValue.APPLICATION_UNSUCCESSFUL,
+        createdByName: 'TE£ST_USER',
+        createdByDateTime: new Date().toISOString(),
+        notes:
+          'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.',
+      },
+    ]
   }
 }
